+++ test/libraries/mathLib/abs.t.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: UNLICENSED
+// SPDX-License-Identifier: MIT
 pragma solidity 0.8.22;
 
 import "test/Base.t.sol";
@@ -10,7 +10,7 @@ contract MathLib_Abs_Test is Base_Test {
 
     function test_abs_negValue(int256 value) external {
         vm.assume(value < 0 && value > (type(int256).min + 1));
-        assertEq(value.abs(), uint(-value));
+        assertEq(value.abs(), uint256(-value));
     }
 
     function test_abs_zero() external {
@@ -20,6 +20,6 @@ contract MathLib_Abs_Test is Base_Test {
 
     function test_abs_posValue(int256 value) external {
         vm.assume(value > 0);
-        assertEq(value.abs(), uint(value));
+        assertEq(value.abs(), uint256(value));
     }
 }
